impl ui

+++ book-network/src/main/java/com/anass/book/feedback/Feedback.java
@@ -20,7 +20,7 @@ import lombok.experimental.SuperBuilder;
 @Entity
 public class Feedback extends BaseEntity {
     @Column
-    private Integer note;
+    private Double note;
     private String comment;
     @ManyToOne
     @JoinColumn(name="book_id")
